requirements.txt
# Browser Automation
playwright==1.41.0
selenium==4.17.2

# AI & NLP
google-generativeai==0.3.2
openai==1.10.0
spacy==3.7.2
PyPDF2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3

# Web Scraping
beautifulsoup4==4.12.3
lxml==5.1.0
requests==2.31.0
httpx==0.26.0

# Data Processing
pandas==2.2.0
numpy==1.26.3

# Configuration & Storage
pyyaml==6.0.1
python-dotenv==1.0.1
sqlalchemy==2.0.25

# Scheduling & Background Tasks
apscheduler==3.10.4
celery==5.3.6

# Email
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1
tqdm==4.66.1
colorama==0.4.6

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-mock==3.12.0

# Security
cryptography==42.0.2
keyring==24.3.0
